Minor critical fixes #28
* Dot movement doesn’t crash when a day view doesn’t hold a reference
to calendar view
* Set -> CVSet
* Some refactoring
